buzzer/requirements.txt
rpi.gpio
redis